refactor: remove duplicate api module
