+++ step-mark-failed.py
@@ -6,7 +6,7 @@
 #
 import json
 import logger
-from api import (
+from module.api import (
     UpdateTaskStatusRequest,
     upload_task_status_update,
 )
